symmetric updates
